+++ packages/devtools_app/lib/src/memory/memory_filter.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:vm_service/vm_service.dart';
 
-import '../flutter_widgets/linked_scroll_controller.dart';
+import '../primitives/linked_scroll_controller.dart';
 import '../primitives/auto_dispose_mixin.dart';
 import '../shared/common_widgets.dart';
 import '../shared/dialogs.dart';
