+++ packages/devtools_app/lib/src/shared/common_widgets.dart
@@ -12,7 +12,7 @@ import 'package:flutter/material.dart';
 
 import '../analytics/analytics.dart' as ga;
 import '../config_specific/launch_url/launch_url.dart';
-import '../flutter_widgets/linked_scroll_controller.dart';
+import '../primitives/linked_scroll_controller.dart';
 import '../primitives/auto_dispose_mixin.dart';
 import '../ui/icons.dart';
 import '../ui/label.dart';
