+++ packages/devtools_app/test/flame_chart_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/src/charts/flame_chart.dart';
-import 'package:devtools_app/src/flutter_widgets/linked_scroll_controller.dart';
+import 'package:devtools_app/src/primitives/linked_scroll_controller.dart';
 import 'package:devtools_app/src/performance/performance_model.dart';
 import 'package:devtools_app/src/profiler/cpu_profile_controller.dart';
 import 'package:devtools_app/src/profiler/cpu_profile_flame_chart.dart';
